(ここchishibukiやんけ…)入力抽象化
