+++ tekiyoke2/Assets/scripts/Input.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c80cd68f28ab8654089163c9b1596fa0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
