+++ tekiyoke2/Assets/scripts/Input/IAskedInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 258c4c28098809b419647ee6c24c820c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
